+++ src/main/java/Duke/Deadline.java
@@ -1,21 +1,21 @@
+package Duke;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 
-public class Event extends Task{
+public class Deadline extends Task{
 
-    protected String at;
-    protected String day;
+    protected String by;
     protected String date;
     protected String time;
     protected String a, b;
 
-    public Event(String name, String at, String date, String time) {
+    public Deadline(String name, String date, String time) {
         super(name);
-        this.at = at;
-        this.time = time;
         this.date = date;
+        this.time = time;
     }
 
     public void dateProcess() {
@@ -23,8 +23,8 @@ public class Event extends Task{
             date = "0" + date;
         }
         LocalDate d1 = LocalDate.parse(date.substring(6, 10) + "-" + date.substring(3, 5) + "-" + date.substring(0, 2));
-        day = d1.getDayOfWeek().toString(); // -> SUNDAY
-        a= d1.format(DateTimeFormatter.ofPattern("MMM dd YYYY")); // -> Dec 27 2001
+        by = d1.getDayOfWeek().toString(); // -> SUNDAY
+        a = d1.format(DateTimeFormatter.ofPattern("MMM dd YYYY")); // -> Dec 27 2001
     }
 
     public void timeProcess() {
@@ -38,6 +38,6 @@ public class Event extends Task{
     public String toString() {
         dateProcess();
         timeProcess();
-        return "[E]" + super.toString() + " (at: " + at + ", " + day + ", " + a + ", " + b + ")";
+        return "[D]" + super.toString() + " (by: " + by + ", " + a + ", " + b + ")";
     }
 }
