+++ src/main/java/Duke/Duke.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.File;
