+++ src/main/java/Duke/Todo.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class Todo extends Task {
 
     public Todo(String name) {
